+++ lib/utils/consts.dart
@@ -1,1 +1,2 @@
 const showBrightSchemeFlagName = 'show_bright_scheme';
+const showThemeConfigFlagName = 'show_theme_config';